Remove assembly calls to runtime.panicindex for Go 1.12
Go 1.12 does not allow the unexported function runtime.panicindex to
be called from assembly code in other packages. Fix this by wrapping
the assembly functions that call panicindex in Go functions and
lifting these checks into the Go functions. The Go wrapper functions
are inlined (as of Go 1.12), so there's no additional cost to doing
this.

Fixes #4.